Working with 3 layers
