+++ src/main/java/ch/schaermedia/ecovolution/Sim.java
@@ -113,6 +113,8 @@ public class Sim extends PApplet {
         PhaseMixture solids = layer.getMixtureForPhase(Phase.SOLID);
         Compound water = solids.getCompound("H2O");
         water.add(10000 * Consts.PRESCISION, 1000 * Consts.PRESCISION);
+        Compound o2 = solids.getCompound("O2");
+        o2.add(10000 * Consts.PRESCISION, 1000 * Consts.PRESCISION);
     }
 
 }
