+++ src/main/java/ch/schaermedia/ecovolution/environment/world/Tile.java
@@ -52,8 +52,16 @@ public class Tile {
         temperature_k = temperatureSum / layers.length;
     }
 
+    public void calculate()
+    {
+        spreadToHigher();
+        importBuffers();
+        spreadToLower();
+        importBuffers();
+        spreadHorizontal();
+    }
 
-    public void spreadHorizontal()
+    private void spreadHorizontal()
     {
         for (LayerMixture layer : layers)
         {
@@ -61,41 +69,48 @@ public class Tile {
         }
     }
 
-    public void importBuffers(){
+    private void importBuffers()
+    {
         for (LayerMixture layer : layers)
         {
             layer.importBuffers();
         }
     }
 
-    public void spreadToHigher()
+    private void spreadToHigher()
     {
         for (int i = 0; i < layers.length - 1; i++)
         {
             LayerMixture layer = layers[i];
+            if (layer.getAmount_mol() == 0)
+            {
+                continue;
+            }
             long molesOverVolume = layer.molesOverVolume();
             if (molesOverVolume == 0)
             {
                 continue;
             }
-            if(molesOverVolume<0){
+            if (molesOverVolume < 0)
+            {
                 throw new RuntimeException("negative moles over volume");
             }
             double percentage = molesOverVolume / layer.getAmount_mol();
-            if (percentage > 1)
+            if (percentage > 1.0)
             {
-                percentage = 1;
+                percentage = 1.0;
             }
             layer.spreadToHigher(percentage);
         }
     }
 
-    public void spreadToLower()
+    private void spreadToLower()
     {
         for (int i = 1; i < layers.length; i++)
         {
             LayerMixture layer = layers[i];
-            if(layer.getAmount_mol() == 0){
+            if (layer.getAmount_mol() == 0)
+            {
                 continue;
             }
             LayerMixture lower = layers[i - 1];
