+++ src/main/java/ch/schaermedia/ecovolution/environment/world/World.java
@@ -5,14 +5,6 @@
  */
 package ch.schaermedia.ecovolution.environment.world;
 
-import java.util.ArrayList;
-import java.util.List;
-import java.util.concurrent.Callable;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
-import java.util.logging.Level;
-import java.util.logging.Logger;
-
 /**
  *
  * @author Quentin
@@ -23,8 +15,6 @@ public class World {
     public static final int NEIGHBOUR_RANGE = 3;
     public static final int NEIGHBOUR_SQUARED = (2 * NEIGHBOUR_RANGE + 1) * (2 * NEIGHBOUR_RANGE + 1);
     public static final double HORIZONTAL_SPREAD_PERCENTAGE = 1.0 / (double) NEIGHBOUR_SQUARED;
-
-    private ExecutorService threadPool;
 
     private Tile[][] grid;
     private final int width;
@@ -37,128 +27,26 @@ public class World {
         this.width = width;
         this.height = height;
         initGrid();
-        threadPool = Executors.newFixedThreadPool(2);
     }
 
     public void update()
     {
-        calculate();
-        updateAll();
+        for (int x = 0; x < width; x++)
+        {
+            for (int y = 0; y < height; y++)
+            {
+                grid[x][y].calculate();
+                amount_mol += grid[x][y].getAmount_mol();
+            }
+        }
         amount_mol = 0;
         for (int x = 0; x < width; x++)
         {
             for (int y = 0; y < height; y++)
             {
+                grid[x][y].update();
                 amount_mol += grid[x][y].getAmount_mol();
             }
-        }
-    }
-
-    private void updateAll(){
-        List<Callable<Object>> tasks = new ArrayList();
-        for (int x = 0; x < width; x++)
-        {
-            for (int y = 0; y < height; y++)
-            {
-                tasks.add(Executors.callable(grid[x][y]::update));
-            }
-        }
-        try
-        {
-            threadPool.invokeAll(tasks);
-        }
-        catch (InterruptedException ex)
-        {
-            Logger.getLogger(World.class.getName()).log(Level.SEVERE, null, ex);
-        }
-    }
-
-    private void calculate()
-    {
-        spreadToHigher();
-        importBuffers();
-        spreadToLower();
-        importBuffers();
-        spreadHorizontal();
-    }
-
-    private void spreadToHigher()
-    {
-        List<Callable<Object>> tasks = new ArrayList();
-        for (int x = 0; x < width; x++)
-        {
-            for (int y = 0; y < height; y++)
-            {
-                tasks.add(Executors.callable(grid[x][y]::spreadToHigher));
-            }
-        }
-        try
-        {
-            threadPool.invokeAll(tasks);
-        }
-        catch (InterruptedException ex)
-        {
-            Logger.getLogger(World.class.getName()).log(Level.SEVERE, null, ex);
-        }
-    }
-
-    private void importBuffers()
-    {
-        List<Callable<Object>> tasks = new ArrayList();
-        for (int x = 0; x < width; x++)
-        {
-            for (int y = 0; y < height; y++)
-            {
-                tasks.add(Executors.callable(grid[x][y]::importBuffers));
-            }
-        }
-        try
-        {
-            threadPool.invokeAll(tasks);
-        }
-        catch (InterruptedException ex)
-        {
-            Logger.getLogger(World.class.getName()).log(Level.SEVERE, null, ex);
-        }
-    }
-
-    private void spreadToLower()
-    {
-        List<Callable<Object>> tasks = new ArrayList();
-        for (int x = 0; x < width; x++)
-        {
-            for (int y = 0; y < height; y++)
-            {
-                tasks.add(Executors.callable(grid[x][y]::spreadToLower));
-            }
-        }
-        try
-        {
-            threadPool.invokeAll(tasks);
-        }
-        catch (InterruptedException ex)
-        {
-            Logger.getLogger(World.class.getName()).log(Level.SEVERE, null, ex);
-        }
-    }
-
-    private void spreadHorizontal()
-    {
-        List<Callable<Object>> tasks = new ArrayList();
-        for (int x = 0; x < width; x++)
-        {
-            for (int y = 0; y < height; y++)
-            {
-                tasks.add(Executors.callable(grid[x][y]::spreadHorizontal));
-            }
-        }
-        try
-        {
-            threadPool.invokeAll(tasks);
-        }
-        catch (InterruptedException ex)
-        {
-            Logger.getLogger(World.class.getName()).log(Level.SEVERE, null, ex);
         }
     }
